feat: отделил массив ссылок от навигации (#73)
