+++ app/features/home/sections/model/const.ts
@@ -30,6 +30,12 @@ export const CARD_LINKS = [
     disable: false,
   },
   {
+    name: 'В разработке',
+    url: '/',
+    img: '/s3/sections/home-link/button_development.webp',
+    disable: true,
+  },
+  {
     name: 'Снаряжение',
     url: '/',
     img: '/s3/sections/home-link/button_items.webp',
@@ -64,42 +70,5 @@ export const CARD_LINKS = [
     url: '/',
     img: '/s3/sections/home-link/button_development.webp',
     disable: true,
-  },
-  {
-    name: 'В разработке',
-    url: '/',
-    img: '/s3/sections/home-link/button_development.webp',
-    disable: true,
-  },
-];
-
-export const SOCIAL_LINKS = [
-  {
-    name: 'Телеграм',
-    icon: 'telegram',
-    url: 'https://t.me/ttgclubnews',
-    disable: false,
-    backgroundColor: '#2798D2',
-  },
-  {
-    name: 'Discord',
-    icon: 'discord',
-    url: 'https://discord.gg/JqFKMKRtxv',
-    disable: false,
-    backgroundColor: '#5865f2',
-  },
-  {
-    name: 'ВКонтакте',
-    icon: 'vk',
-    url: 'https://vk.com/ttg.club',
-    disable: false,
-    backgroundColor: '#097BFF',
-  },
-  {
-    name: 'Boosty',
-    icon: 'boosty',
-    url: 'https://boosty.to/dnd5club',
-    disable: false,
-    backgroundColor: '#F6632B',
   },
 ];
