+++ app/features/home/sections/model/index.ts
@@ -1,1 +1,1 @@
-export { CARD_LINKS, SOCIAL_LINKS } from './const';
+export { CARD_LINKS } from './const';
