+++ app/features/home/social-links/model/const.ts
@@ -1,0 +1,30 @@
+export const SOCIAL_LINKS = [
+  {
+    name: 'Телеграм',
+    icon: 'telegram',
+    url: 'https://t.me/ttgclubnews',
+    disable: false,
+    backgroundColor: '#2798D2',
+  },
+  {
+    name: 'Discord',
+    icon: 'discord',
+    url: 'https://discord.gg/JqFKMKRtxv',
+    disable: false,
+    backgroundColor: '#5865f2',
+  },
+  {
+    name: 'ВКонтакте',
+    icon: 'vk',
+    url: 'https://vk.com/ttg.club',
+    disable: false,
+    backgroundColor: '#097BFF',
+  },
+  {
+    name: 'Boosty',
+    icon: 'boosty',
+    url: 'https://boosty.to/dnd5club',
+    disable: false,
+    backgroundColor: '#F6632B',
+  },
+];
