+++ app/features/home/social-links/model/index.ts
@@ -1,0 +1,1 @@
+export { SOCIAL_LINKS } from './const';
